Pass list of trending posts in Page/Main.js
