+++ src/Constants/TrendingPosts.js
@@ -1,0 +1,40 @@
+const TrendingPosts = [
+  {
+    title: "title",
+    writier: "writer",
+    data: "date",
+    time: "time",
+  },
+  {
+    title: "title",
+    writier: "writer",
+    data: "date",
+    time: "time",
+  },
+  {
+    title: "title",
+    writier: "writer",
+    data: "date",
+    time: "time",
+  },
+  {
+    title: "title",
+    writier: "writer",
+    data: "date",
+    time: "time",
+  },
+  {
+    title: "title",
+    writier: "writer",
+    data: "date",
+    time: "time",
+  },
+  {
+    title: "title",
+    writier: "writer",
+    data: "date",
+    time: "time",
+  },
+]
+
+export default TrendingPosts
