+++ src/Pages/Main/Main.js
@@ -1,7 +1,8 @@
 import { Main } from "../../Components/Main"
+import TrendingPosts from "../../Constants/TrendingPosts"
 
 const MainPage = () => {
-  return <Main />
+  return <Main TrendingPosts={TrendingPosts} />
 }
 
 export default MainPage
